Applying aggregates, utilizing encapsulation
