+++ COVID19Py/analysis.py
@@ -1,0 +1,144 @@
+from typing import Dict, List
+import requests
+import json
+
+class Analysis:
+    default_url = "https://covid-tracker-us.herokuapp.com"
+    url = ""
+    data_source = ""
+    previousData = None
+    latestData = None
+    _valid_data_sources = []
+
+    mirrors_source = "https://raw.github.com/Kamaropoulos/COVID19Py/master/mirrors.json"
+    mirrors = None
+
+    def __init__(self, url="https://covid-tracker-us.herokuapp.com", data_source='jhu'):
+        # Skip mirror checking if custom url was passed
+        if url == self.default_url:
+            # Load mirrors
+            response = requests.get(self.mirrors_source)
+            response.raise_for_status()
+            self.mirrors = response.json()
+
+            # Try to get sources as a test
+            for mirror in self.mirrors:
+                # Set URL of mirror
+                self.url = mirror["url"]
+                result = None
+                try:
+                    result = self._getSources()
+                except Exception as e:
+                    # URL did not work, reset it and move on
+                    self.url = ""
+                    continue
+
+                # TODO: Should have a better health-check, this is way too hacky...
+                if "jhu" in result:
+                    # We found a mirror that worked just fine, let's stick with it
+                    break
+
+                # None of the mirrors worked. Raise an error to inform the user.
+                raise RuntimeError("No available API mirror was found.")
+
+        else:
+            self.url = url
+
+        self._valid_data_sources = self._getSources()
+        if data_source not in self._valid_data_sources:
+            raise ValueError("Invalid data source. Expected one of: %s" % self._valid_data_sources)
+        self.data_source = data_source
+
+    def _update(self, timelines):
+        latest = self.getLatest()
+        locations = self.getLocations(timelines)
+        if self.latestData:
+            self.previousData = self.latestData
+        self.latestData = {
+            "latest": latest,
+            "locations": locations
+        }
+
+    def _getSources(self):
+        response = requests.get(self.url + "/v2/sources")
+        response.raise_for_status()
+        return response.json()["sources"]
+
+    def _request(self, endpoint, params=None):
+        if params is None:
+            params = {}
+        response = requests.get(self.url + endpoint, {**params, "source":self.data_source})
+        response.raise_for_status()
+        return response.json()
+
+    def getAll(self, timelines=False):
+        self._update(timelines)
+        return self.latestData
+
+    def getLatestChanges(self):
+        changes = None
+        if self.previousData:
+            changes = {
+                "confirmed": self.latestData["latest"]["confirmed"] - self.latestData["latest"]["confirmed"],
+                "deaths": self.latestData["latest"]["deaths"] - self.latestData["latest"]["deaths"],
+                "recovered": self.latestData["latest"]["recovered"] - self.latestData["latest"]["recovered"],
+            }
+        else:
+            changes = {
+                "confirmed": 0,
+                "deaths": 0,
+                "recovered": 0,
+            }
+        return changes
+
+    def getLatest(self) -> List[Dict[str, int]]:
+
+        data = self._request("/v2/latest")
+        return data["latest"]
+
+    def getLocations(self, timelines=False, rank_by: str = None) -> List[Dict]:
+
+        data = None
+        if timelines:
+            data = self._request("/v2/locations", {"timelines": str(timelines).lower()})
+        else:
+            data = self._request("/v2/locations")
+
+        data = data["locations"]
+        
+        ranking_criteria = ['confirmed', 'deaths', 'recovered']
+        if rank_by is not None:
+            if rank_by not in ranking_criteria:
+                raise ValueError("Invalid ranking criteria. Expected one of: %s" % ranking_criteria)
+
+            ranked = sorted(data, key=lambda i: i['latest'][rank_by], reverse=True)
+            data = ranked
+
+        return data
+
+    def getLocationByCountryCode(self, country_code, timelines=False) -> List[Dict]:
+
+        data = None
+        if timelines:
+            data = self._request("/v2/locations", {"country_code": country_code, "timelines": str(timelines).lower()})
+        else:
+            data = self._request("/v2/locations", {"country_code": country_code})
+        return data["locations"]
+    
+    def getLocationByCountry(self, country, timelines=False) -> List[Dict]:
+
+        data = None
+        if timelines:
+            data = self._request("/v2/locations", {"country": country, "timelines": str(timelines).lower()})
+        else:
+            data = self._request("/v2/locations", {"country": country})
+        return data["locations"]
+
+    def getLocationById(self, country_id: int):
+
+        data = self._request("/v2/locations/" + str(country_id))
+        return data["location"]
+
+
+
+
